Refactor crime and plot
Several attributes were associated with the wrong objects
making things awkward. Population weighting is handled by
the crime object, and the plot object does not make changes
to the data. Choosing which column to plot is handled entirely
by the plot object. As this involves the time, the title and name
are now handled by the plot object as well. Time filtering in the
Crime class may need to be looked at to make this work.